fleet/state.py: Add DispatchRecord.from_dict and decompose into 4 files
P9-F1: Add DispatchRecord.from_dict() classmethod that encapsulates
legacy l2_*/l3_* field-name fallback chains, eliminating duplicated
.get() calls and centralizing deserialization logic.

P8-F4: Extract state.py into four files:
- state_types.py: types + constants (DispatchStatus, DispatchRecord,
  CampaignState, ResumeDecision, GateRecordResult, transition rules)
- state_gates.py: record_gate_outcome
- state_recovery.py: crash_recover_dispatch, resume_campaign_from_state,
  has_failed_dispatch
- state.py (slimmed): I/O + mutations + re-exports facade

state.py re-exports all public symbols so existing import paths remain
valid. state_gates.py and state_recovery.py use deferred imports for
state.py symbols to avoid circular imports.
